requirements.txt
torch
facenet-pytorch
scikit-learn
numpy
Flask
flask-cors
gunicorn
python-dotenv
Pillow
joblib
tqdm
